+++ changelog.md
@@ -3,6 +3,12 @@
 
 > **Note:** The dates included here are the dates of the commits that were made to the repository. The actual release dates may be different.
 
+## 14 October 2024 (dev)
+* Add the `context` tab to provide test data to the pre-populate and extract operations in the questionnaire tester.
+  * Supports `subject` and `author` as parameters to the tester via the URL (If you want to include the display as well as the reference, put it after a comma in the URL parameter e.g. `subject=Patient/123,John Doe`) 
+* Tidy up the tab numbering in the questionnaire tester so that it is selected by name rather than number (to easy maintenance)
+* Update the old "pre-populate" tab to show the items (missed passing through the collection)
+
 ## 10 October 2024 (dev)
 * Models tab in the form tester will auto-load any model query passed in the URL
 * Extract tab layout cleaned up